Assignment 6: After Tutorial
Replaced "set" with "upd"

Set will replace the entire SDS with a new value (which is a modified copy), which might overwrite someone else's work if that person modified the SDS between the bind and the set.
Update will maintain the original SDS and just add the new info (if the update function does just that). Therefore, it won't affect other's people work and it's safer.